% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesamp.R
\name{onesampb.ccmc}
\alias{onesampb.ccmc}
\title{One sample percentile bootstrap test with cluster correction for multiple comparisons}
\usage{
onesampb.ccmc(x, nboot = 599)
}
\description{
One sample percentile bootstrap test with cluster correction for multiple comparisons
}
