% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesamp.R
\name{trimbt.ccmc}
\alias{trimbt.ccmc}
\title{One sample t-test on trimmed means with cluster correction for multiple comparisons}
\usage{
trimbt.ccmc(x, tr = 0.2, bt = FALSE, nboot = 599)
}
\arguments{
\item{x}{A list or a matrix. In the first case
x[[1]] contains the data for the first group, x[[2]] the data
for the second group, etc. Length(x) = the number of groups = J.
If stored in a matrix, columns correspond to groups.

[[1]: R:[1
[[2]: R:[2}

\item{tr}{The amount of trimming - default 0.2. Set tr=0 for a t-test on means.}

\item{bt}{Set to TRUE to use critical t values based on boostrap-t distributions - default = FALSE}
}
\value{
A list of univariate t values, p values and confidence intervals, as well as cluster-based statistics.
}
\description{
One sample t-test on trimmed means with cluster correction for multiple comparisons
}
\section{References}{
NA
}

\seealso{
\code{}
}
