% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{trimparts}
\alias{trimparts}
\title{Compute the trimmed mean, effective sample size, and squared standard error.
The default amount of trimming is tr=.2.}
\usage{
trimparts(x, tr = 0.2)
}
\description{
Compute the trimmed mean, effective sample size, and squared standard error.
The default amount of trimming is tr=.2.
}
