% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{trimse}
\alias{trimse}
\title{Estimate the standard error of the trimmed mean
The default amount of trimming is tr=.2.}
\usage{
trimse(x, tr = 0.2, na.rm = FALSE)
}
\description{
Estimate the standard error of the trimmed mean
The default amount of trimming is tr=.2.
}
