% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{winvar}
\alias{winvar}
\title{Compute the Winsorized variance for the data in the vector x.
tr is the amount of Winsorization which defaults to .2.}
\usage{
winvar(x, tr = 0.2, na.rm = FALSE, STAND = NULL)
}
\description{
Compute the Winsorized variance for the data in the vector x.
tr is the amount of Winsorization which defaults to .2.
}
